refactor: create subpackages for DTOs
