+++ src/main/java/com/example/mercadofrescos/controller/InboundOrderController.java
@@ -1,7 +1,7 @@
 package com.example.mercadofrescos.controller;
 
-import com.example.mercadofrescos.dto.InboundOrderResponseDTO;
-import com.example.mercadofrescos.dto.InsertBatchRequestDTO;
+import com.example.mercadofrescos.dto.inboundOrder.InboundOrderResponseDTO;
+import com.example.mercadofrescos.dto.inboundOrder.InsertBatchRequestDTO;
 import com.example.mercadofrescos.model.InboundOrder;
 import com.example.mercadofrescos.service.InboundOrderService;
 import lombok.RequiredArgsConstructor;
