+++ src/main/java/com/example/mercadofrescos/controller/PurchaseOrderController.java
@@ -1,9 +1,10 @@
 package com.example.mercadofrescos.controller;
 
-import com.example.mercadofrescos.dto.*;
-import com.example.mercadofrescos.model.enums.OrderBy;
+import com.example.mercadofrescos.dto.purchase.PurchaseItemResponseDTO;
+import com.example.mercadofrescos.dto.purchase.PurchaseOrderRequestDTO;
+import com.example.mercadofrescos.dto.purchase.PurchasePriceDTO;
+import com.example.mercadofrescos.dto.purchase.StatusOrderDTO;
 import com.example.mercadofrescos.model.enums.StatusOrder;
-import com.example.mercadofrescos.service.interfaces.IBatchStockService;
 import com.example.mercadofrescos.service.interfaces.IPurchaseOrderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
