+++ src/main/java/com/example/mercadofrescos/dto/batchStock/BatchStockDTO.java
@@ -1,4 +1,4 @@
-package com.example.mercadofrescos.dto;
+package com.example.mercadofrescos.dto.batchStock;
 
 import com.example.mercadofrescos.model.BatchStock;
 import com.example.mercadofrescos.model.Product;
