+++ src/main/java/com/example/mercadofrescos/dto/product/ProductDTO.java
@@ -1,4 +1,4 @@
-package com.example.mercadofrescos.dto;
+package com.example.mercadofrescos.dto.product;
 
 import com.example.mercadofrescos.model.enums.Category;
 import lombok.AllArgsConstructor;
