+++ src/main/java/com/example/mercadofrescos/dto/purchase/PurchaseOrderDTO.java
@@ -1,4 +1,4 @@
-package com.example.mercadofrescos.dto;
+package com.example.mercadofrescos.dto.purchase;
 
 import com.example.mercadofrescos.model.PurchaseItem;
 import com.fasterxml.jackson.annotation.JsonProperty;
