+++ src/main/java/com/example/mercadofrescos/service/interfaces/IBatchStockService.java
@@ -1,9 +1,8 @@
 package com.example.mercadofrescos.service.interfaces;
 
-import com.example.mercadofrescos.dto.BatchStockResponseDTO;
+import com.example.mercadofrescos.dto.batchStock.BatchStockResponseDTO;
 import com.example.mercadofrescos.model.BatchStock;
 import com.example.mercadofrescos.model.InboundOrder;
-import com.example.mercadofrescos.model.enums.Category;
 import com.example.mercadofrescos.model.enums.OrderBy;
 
 import java.util.List;
