+++ src/main/java/com/example/mercadofrescos/service/interfaces/IPurchaseOrderService.java
@@ -1,8 +1,8 @@
 package com.example.mercadofrescos.service.interfaces;
 
-import com.example.mercadofrescos.dto.PurchaseItemResponseDTO;
-import com.example.mercadofrescos.dto.PurchaseOrderRequestDTO;
-import com.example.mercadofrescos.dto.PurchasePriceDTO;
+import com.example.mercadofrescos.dto.purchase.PurchaseItemResponseDTO;
+import com.example.mercadofrescos.dto.purchase.PurchaseOrderRequestDTO;
+import com.example.mercadofrescos.dto.purchase.PurchasePriceDTO;
 import com.example.mercadofrescos.model.PurchaseOrder;
 import com.example.mercadofrescos.model.enums.StatusOrder;
