+++ src/main/java/com/example/mercadofrescos/service/interfaces/IWarehouseService.java
@@ -1,7 +1,6 @@
 package com.example.mercadofrescos.service.interfaces;
 
-import com.example.mercadofrescos.dto.ProductWarehousesDTO;
-import com.example.mercadofrescos.dto.WarehouseProductsDTO;
+import com.example.mercadofrescos.dto.product.ProductWarehousesDTO;
 import com.example.mercadofrescos.model.Warehouse;
 
 public interface IWarehouseService {
