+++ src/test/java/com/example/mercadofrescos/service/ProductServiceTest.java
@@ -1,9 +1,9 @@
 package com.example.mercadofrescos.service;
 
-import com.example.mercadofrescos.dto.BatchStockAgentResponseDTO;
-import com.example.mercadofrescos.dto.ProductAgentResponseDTO;
-import com.example.mercadofrescos.dto.ProductDTO;
-import com.example.mercadofrescos.dto.ProductResponseDTO;
+import com.example.mercadofrescos.dto.batchStock.BatchStockAgentResponseDTO;
+import com.example.mercadofrescos.dto.product.ProductAgentResponseDTO;
+import com.example.mercadofrescos.dto.product.ProductDTO;
+import com.example.mercadofrescos.dto.product.ProductResponseDTO;
 import com.example.mercadofrescos.exception.InvalidQueryParamException;
 import com.example.mercadofrescos.exception.NotFoundException;
 import com.example.mercadofrescos.exception.ProductsListNotFoundException;
