+++ src/test/java/com/example/mercadofrescos/service/PurchaseOrderTest.java
@@ -1,9 +1,9 @@
 package com.example.mercadofrescos.service;
 
 
-import com.example.mercadofrescos.dto.PurchaseItemResponseDTO;
-import com.example.mercadofrescos.dto.PurchaseOrderRequestDTO;
-import com.example.mercadofrescos.dto.PurchasePriceDTO;
+import com.example.mercadofrescos.dto.purchase.PurchaseItemResponseDTO;
+import com.example.mercadofrescos.dto.purchase.PurchaseOrderRequestDTO;
+import com.example.mercadofrescos.dto.purchase.PurchasePriceDTO;
 import com.example.mercadofrescos.exception.InvalidPurchaseException;
 import com.example.mercadofrescos.exception.NotFoundException;
 import com.example.mercadofrescos.mocks.*;
